feat(coach): merge dashboard test into main coach dashboard and fix login redirect test
